Rename dogs component to dogs-list
