+++ src/app/modules/dogs/dogs.module.ts
@@ -1,6 +1,6 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
-import { DogsComponent } from './components/dogs/dogs.component';
+import { DogsComponent } from './components/dogs-list/dogs-list.component';
 import { DogsStoreModule } from './store/dogs-store.module';
 import { DogsService } from './services/dogs.service';
 import { DogsApiService } from './services/dogs-api.service';
